Extract reusable big query data conversion concern
