+++ app/models/concerns/entity_events.rb
@@ -2,15 +2,17 @@ module EntityEvents
   extend ActiveSupport::Concern
 
   included do
+    include BigqueryDataConversion
+
     attr_accessor :event_tags
 
     after_create do
-      data = entity_data(attributes)
+      data = entity_data(attributes, self.class.table_name)
       send_event('create_entity', data) if data.any?
     end
 
     after_destroy do
-      data = entity_data(attributes)
+      data = entity_data(attributes, self.class.table_name)
       send_event('delete_entity', data) if data.any?
     end
 
@@ -18,10 +20,10 @@ module EntityEvents
       # in this after_update hook we don’t have access to the new fields via
       # #attributes — we need to dig them out of saved_changes which stores
       # them in the format { attr: ['old', 'new'] }
-      interesting_changes = entity_data(saved_changes.transform_values(&:last))
+      interesting_changes = entity_data(saved_changes.transform_values(&:last), self.class.table_name)
 
       if interesting_changes.any?
-        send_event('update_entity', entity_data(attributes).merge(interesting_changes))
+        send_event('update_entity', entity_data(attributes, self.class.table_name).merge(interesting_changes))
       end
     end
   end
@@ -37,20 +39,5 @@ module EntityEvents
       .with_request_uuid(RequestLocals.fetch(:request_id) { nil })
 
     SendEventsToBigquery.perform_async([event.as_json])
-  end
-
-  def entity_data(changeset)
-    exportable_attrs = Rails.configuration.analytics[self.class.table_name.to_sym].presence || []
-    pii_attrs = Rails.configuration.analytics_pii[self.class.table_name.to_sym].presence || []
-    exportable_pii_attrs = exportable_attrs & pii_attrs
-
-    to_send = changeset.slice(*exportable_attrs&.map(&:to_s))
-    to_obfuscate = changeset.slice(*exportable_pii_attrs&.map(&:to_s))
-
-    to_send.deep_merge(to_obfuscate.transform_values { |value| anonymise(value) })
-  end
-
-  def anonymise(value)
-    Digest::SHA2.hexdigest(value.to_s)
   end
 end
